+++ pkg/deploy/kubernetes_parse.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package fl_k8s
+package deploy
 
 import (
 	"strings"
